+++ pkg/azurecluster/mutate_update.go
@@ -8,9 +8,11 @@ import (
 	"github.com/giantswarm/micrologger"
 	"k8s.io/api/admission/v1beta1"
 	capz "sigs.k8s.io/cluster-api-provider-azure/api/v1alpha3"
+	capzv1alpha3 "sigs.k8s.io/cluster-api-provider-azure/api/v1alpha3"
 	ctrl "sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/giantswarm/azure-admission-controller/pkg/generic"
+	"github.com/giantswarm/azure-admission-controller/pkg/key"
 	"github.com/giantswarm/azure-admission-controller/pkg/mutator"
 )
 
@@ -53,7 +55,47 @@ func (m *UpdateMutator) Mutate(ctx context.Context, request *v1beta1.AdmissionRe
 		return []mutator.PatchOperation{}, microerror.Maskf(parsingFailedError, "unable to parse AzureCluster CR: %v", err)
 	}
 
-	patch, err := generic.EnsureComponentVersionLabelFromRelease(ctx, m.ctrlClient, azureClusterCR.GetObjectMeta(), "azure-operator", label.AzureOperatorVersion)
+	patch, err := m.ensureAPIServerLBName(ctx, azureClusterCR)
+	if err != nil {
+		return []mutator.PatchOperation{}, microerror.Mask(err)
+	}
+	if patch != nil {
+		result = append(result, *patch)
+	}
+
+	patch, err = m.ensureAPIServerLBSKU(ctx, azureClusterCR)
+	if err != nil {
+		return []mutator.PatchOperation{}, microerror.Mask(err)
+	}
+	if patch != nil {
+		result = append(result, *patch)
+	}
+
+	patch, err = m.ensureAPIServerLBType(ctx, azureClusterCR)
+	if err != nil {
+		return []mutator.PatchOperation{}, microerror.Mask(err)
+	}
+	if patch != nil {
+		result = append(result, *patch)
+	}
+
+	patch, err = m.ensureAPIServerLBFrontendIPs(ctx, azureClusterCR)
+	if err != nil {
+		return []mutator.PatchOperation{}, microerror.Mask(err)
+	}
+	if patch != nil {
+		result = append(result, *patch)
+	}
+
+	patch, err = m.ensureControlPlaneSubnet(ctx, azureClusterCR)
+	if err != nil {
+		return []mutator.PatchOperation{}, microerror.Mask(err)
+	}
+	if patch != nil {
+		result = append(result, *patch)
+	}
+
+	patch, err = generic.EnsureComponentVersionLabelFromRelease(ctx, m.ctrlClient, azureClusterCR.GetObjectMeta(), "azure-operator", label.AzureOperatorVersion)
 	if err != nil {
 		return []mutator.PatchOperation{}, microerror.Mask(err)
 	}
@@ -78,4 +120,62 @@ func (m *UpdateMutator) Log(keyVals ...interface{}) {
 
 func (m *UpdateMutator) Resource() string {
 	return "azurecluster"
+}
+
+func (m *UpdateMutator) ensureAPIServerLBName(ctx context.Context, azureCluster *capzv1alpha3.AzureCluster) (*mutator.PatchOperation, error) {
+	if len(azureCluster.Spec.NetworkSpec.APIServerLB.Name) < 1 {
+		return mutator.PatchAdd("/spec/networkSpec/apiServerLB/name", key.APIServerLBName(azureCluster.Name)), nil
+	}
+
+	return nil, nil
+}
+
+func (m *UpdateMutator) ensureAPIServerLBSKU(ctx context.Context, azureCluster *capzv1alpha3.AzureCluster) (*mutator.PatchOperation, error) {
+	if len(azureCluster.Spec.NetworkSpec.APIServerLB.SKU) < 1 {
+		return mutator.PatchAdd("/spec/networkSpec/apiServerLB/sku", key.APIServerLBSKU()), nil
+	}
+
+	return nil, nil
+}
+
+func (m *UpdateMutator) ensureAPIServerLBType(ctx context.Context, azureCluster *capzv1alpha3.AzureCluster) (*mutator.PatchOperation, error) {
+	if len(azureCluster.Spec.NetworkSpec.APIServerLB.Type) < 1 {
+		return mutator.PatchAdd("/spec/networkSpec/apiServerLB/type", key.APIServerLBType()), nil
+	}
+
+	return nil, nil
+}
+
+func (m *UpdateMutator) ensureAPIServerLBFrontendIPs(ctx context.Context, azureCluster *capzv1alpha3.AzureCluster) (*mutator.PatchOperation, error) {
+	if len(azureCluster.Spec.NetworkSpec.APIServerLB.FrontendIPs) < 1 {
+		frontendIPs := []capzv1alpha3.FrontendIP{
+			{Name: key.APIServerLBFrontendIPName(azureCluster.Name)},
+		}
+
+		return mutator.PatchAdd("/spec/networkSpec/apiServerLB/frontendIPs", frontendIPs), nil
+	}
+
+	return nil, nil
+}
+
+func (m *UpdateMutator) ensureControlPlaneSubnet(ctx context.Context, azureCluster *capzv1alpha3.AzureCluster) (*mutator.PatchOperation, error) {
+	hasControlPlaneSubnet := false
+	for _, subnet := range azureCluster.Spec.NetworkSpec.Subnets {
+		if subnet.Role == capzv1alpha3.SubnetControlPlane {
+			hasControlPlaneSubnet = true
+			break
+		}
+	}
+
+	if !hasControlPlaneSubnet {
+		subnets := azureCluster.Spec.NetworkSpec.Subnets[:]
+		subnets = append(subnets, &capzv1alpha3.SubnetSpec{
+			Role: capzv1alpha3.SubnetControlPlane,
+			Name: key.MasterSubnetName(azureCluster.Name),
+		})
+
+		return mutator.PatchAdd("/spec/networkSpec/subnets", subnets), nil
+	}
+
+	return nil, nil
 }