+++ pkg/azurecluster/validate_update_test.go
@@ -31,7 +31,7 @@ func TestAzureClusterUpdateValidate(t *testing.T) {
 		{
 			name:            "case 1: host changed",
 			oldAzureCluster: builder.BuildAzureClusterAsJson(builder.Name("ab123"), builder.ControlPlaneEndpoint("api.ab123.k8s.test.westeurope.azure.gigantic.io", 443)),
-			newAzureCluster: builder.BuildAzureClusterAsJson(builder.ControlPlaneEndpoint("api.azure.gigantic.io", 443)),
+			newAzureCluster: builder.BuildAzureClusterAsJson(builder.Name("ab123"), builder.ControlPlaneEndpoint("api.azure.gigantic.io", 443)),
 			errorMatcher:    IsControlPlaneEndpointWasChangedError,
 		},
 		{
@@ -39,12 +39,6 @@ func TestAzureClusterUpdateValidate(t *testing.T) {
 			oldAzureCluster: builder.BuildAzureClusterAsJson(builder.Name("ab123"), builder.ControlPlaneEndpoint("api.ab123.k8s.test.westeurope.azure.gigantic.io", 443)),
 			newAzureCluster: builder.BuildAzureClusterAsJson(builder.Name("ab123"), builder.ControlPlaneEndpoint("api.ab123.k8s.test.westeurope.azure.gigantic.io", 80)),
 			errorMatcher:    IsControlPlaneEndpointWasChangedError,
-		},
-		{
-			name:            "case 3: location changed",
-			oldAzureCluster: builder.BuildAzureClusterAsJson(builder.Name("ab123"), builder.Location("westeurope")),
-			newAzureCluster: builder.BuildAzureClusterAsJson(builder.Name("ab123"), builder.Location("westpoland")),
-			errorMatcher:    IsLocationWasChangedError,
 		},
 	}