+++ pkg/azuremachine/validate_create.go
@@ -59,7 +59,13 @@ func (a *CreateValidator) Validate(ctx context.Context, request *v1beta1.Admissi
 		return microerror.Maskf(errors.ParsingFailedError, "unable to parse AzureMachine CR: %v", err)
 	}
 
-	err := generic.ValidateOrganizationLabelContainsExistingOrganization(ctx, a.ctrlClient, cr)
+	err := cr.ValidateCreate()
+	err = errors.IgnoreCAPIErrorForField("sshPublicKey", err)
+	if err != nil {
+		return microerror.Mask(err)
+	}
+
+	err = generic.ValidateOrganizationLabelContainsExistingOrganization(ctx, a.ctrlClient, cr)
 	if err != nil {
 		return microerror.Mask(err)
 	}