+++ pkg/azuremachine/validate_update.go
@@ -52,7 +52,13 @@ func (a *UpdateValidator) Validate(ctx context.Context, request *v1beta1.Admissi
 		return microerror.Maskf(errors.ParsingFailedError, "unable to parse AzureMachine CR: %v", err)
 	}
 
-	err := checkSSHKeyIsEmpty(ctx, azureMachineNewCR)
+	err := azureMachineNewCR.ValidateUpdate(azureMachineOldCR)
+	err = errors.IgnoreCAPIErrorForField("sshPublicKey", err)
+	if err != nil {
+		return microerror.Mask(err)
+	}
+
+	err = checkSSHKeyIsEmpty(ctx, azureMachineNewCR)
 	if err != nil {
 		return microerror.Mask(err)
 	}