+++ pkg/key/key.go
@@ -1,6 +1,8 @@
 package key
 
-import "fmt"
+import (
+	"fmt"
+)
 
 const (
 	ControlPlaneEndpointPort  = 443
@@ -13,4 +15,28 @@ func GetControlPlaneEndpointHost(clusterName string, baseDomain string) string {
 
 func ServiceDomain() string {
 	return "cluster.local"
+}
+
+func APIServerLBName(clusterName string) string {
+	return fmt.Sprintf("%s-%s-%s", clusterName, "API", "PublicLoadBalancer")
+}
+
+func APIServerLBSKU() string {
+	return "Standard"
+}
+
+func APIServerLBType() string {
+	return "Public"
+}
+
+func APIServerLBFrontendIPName(clusterName string) string {
+	return fmt.Sprintf("%s-%s", APIServerLBName(clusterName), "Frontend")
+}
+
+func OSDiskCachingType() string {
+	return "ReadWrite"
+}
+
+func MasterSubnetName(clusterName string) string {
+	return fmt.Sprintf("%s-%s-%s", clusterName, "VirtualNetwork", "MasterSubnet")
 }