+++ pkg/machinepool/validate_machinepool_update.go
@@ -43,7 +43,12 @@ func (a *UpdateValidator) Validate(ctx context.Context, request *v1beta1.Admissi
 		return microerror.Maskf(parsingFailedError, "unable to parse machinePool CR: %v", err)
 	}
 
-	err := generic.ValidateOrganizationLabelUnchanged(machinePoolOldCR, machinePoolNewCR)
+	err := machinePoolNewCR.ValidateUpdate(machinePoolOldCR)
+	if err != nil {
+		return microerror.Mask(err)
+	}
+
+	err = generic.ValidateOrganizationLabelUnchanged(machinePoolOldCR, machinePoolNewCR)
 	if err != nil {
 		return microerror.Mask(err)
 	}