+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Execute CAPZ validation for all resources.
+
 ## [2.2.0] - 2021-02-05
 
 ### Fixed
@@ -14,7 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `Cluster` and `AzureCluster` mutate webhook definition in the Helm chart.
 - Ensure `cluster-operator.giantswarm.io/version` label has the right value depending on the `release.giantswarm.io/version`
   label when updating `Cluster` and `AzureCluster`.
-  
+
 ## [2.1.0] - 2021-02-03
 
 ### Added
@@ -76,7 +80,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Default `Cluster.Spec.ClusterNetwork.ServiceDomain` to `cluster.local` and don't allow any other value to be set. 
+- Default `Cluster.Spec.ClusterNetwork.ServiceDomain` to `cluster.local` and don't allow any other value to be set.
 
 ## [1.13.2] - 2020-11-13
 
@@ -118,7 +122,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Prevent Organization label value change on CR updates. 
+- Prevent Organization label value change on CR updates.
 
 ### Removed